tests/wpt/css-tests/css-text-3_dev/html/seg-break-transformation-001.htm

<!DOCTYPE html>
<html lang="en"><head>
<meta charset="utf-8">
<title>Wide characters around line break</title>
<link href="mailto:ishida@w3.org" rel="author" title="Richard Ishida">
<link href="https://drafts.csswg.org/css-text-3/#line-break-transform" rel="help">
<meta content="If the East Asian Width property of both the character before and after the line feed is W and neither side is Hangul, then the segment break is removed." name="assert">
<style type="text/css">
/* the CSS below is not part of the test */
.test span { font-size: 24px; font-family: sans-serif; background-color:#2AA5F7; color: white; }
.ref span { font-size: 24px; font-family: sans-serif; background-color:#270CEF; color: white; }
</style>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div id="log"></div>
<div class="test" id="test1"><span>日本語
中国话</span></div>
<div class="test" id="test2"><span>日本語   
中国话</span></div>
<div class="test" id="test3"><span>日本語
        中国话</span></div>
<div class="test" id="test4"><span>日本語   
     中国话</span></div>
<div class="test" id="test5"><span>日本語


中国话</span></div>
<div class="test" id="test6"><span>日本語  
   
   
   中国话</span></div>
<div class="ref" id="ref"><span>日本語中国话</span></div>
<script>
test(function() {
assert_equals(document.getElementById('test1').firstChild.offsetWidth, document.getElementById('ref').firstChild.offsetWidth);
}, "linebreak only");
test(function() {
assert_equals(document.getElementById('test2').firstChild.offsetWidth, document.getElementById('ref').firstChild.offsetWidth);
}, "spaces linebreak");
test(function() {
assert_equals(document.getElementById('test3').firstChild.offsetWidth, document.getElementById('ref').firstChild.offsetWidth);
}, "linebreak spaces");
test(function() {
assert_equals(document.getElementById('test4').firstChild.offsetWidth, document.getElementById('ref').firstChild.offsetWidth);
}, "spaces linebreak spaces");
test(function() {
assert_equals(document.getElementById('test5').firstChild.offsetWidth, document.getElementById('ref').firstChild.offsetWidth);
}, "multiple linebreaks");
test(function() {
assert_equals(document.getElementById('test6').firstChild.offsetWidth, document.getElementById('ref').firstChild.offsetWidth);
}, "multiple linebreaks + spaces");
</script>
<!-- Notes:
The assertion will fail if space is produced for any line in the test paragraph.
-->




</body></html>
